* added command-line options w/selection of actions * improved progress reporting

+++ build/loading/bin/update_tax_ids.pl
@@ -10,28 +10,69 @@
 use strict;
 use warnings;
 use Unison;
+use Getopt::Long qw(:config gnu_getopt);
+use Data::Dumper;
 
-my $palias_id_inc = 10000;
-my $palias_id_min = 0;
-my $palias_id_max = 10000000;
+my %opts =
+  (
+   'increment' => 1000,
+   'update-increment' => 100,
+   'palias-id-min' => 0,
+   'palias-id-max' => undef,
+
+   'drop-paotax' => 0,
+   'create-paotax' => 0,
+   'assign' => 0,
+   'reassign' => 0,
+   'nullify' => 0
+  );
+
+
+GetOptions(\%opts,
+		   'increment|i=i',
+		   'update-increment|u=i',
+		   'palias-id-min|b=i',
+		   'palias-id-max|e=i',
+
+		   'drop-paotax|D+',
+		   'create-paotax|c+',
+		   'assign|a+',
+		   'reassign|r+',
+		   'nullify|n+',
+
+		   'all|A' => sub { $opts{assign} = $opts{reassign} = $opts{nullify} = 1 }
+		   )
+  || die("$0: aye, you've got usage issues, mate\n");
+
 
 sub create_paotax ($);
 sub assign_tax_ids ($);
 sub reassign_tax_ids ($);
 sub nullify_tax_ids ($);
 
+select(STDERR); $|++;
+select(STDOUT); $|++;
+
 
 my $u = new Unison;
 
 
-select(STDERR); $|++;
-select(STDOUT); $|++;
+if (not defined $opts{'palias-id-max'}) {
+  $opts{'palias-id-max'} = $u->selectrow_array('select max(palias_id) from paliasorigin');
+}
 
-create_paotax($u);
-assign_tax_ids($u);
-reassign_tax_ids($u);
-nullify_tax_ids($u);
 
+
+create_paotax($u) 						if $opts{'create-paotax'};
+
+my $nrows_changed = 0;
+$nrows_changed += assign_tax_ids($u)	if $opts{'assign'};
+$nrows_changed += reassign_tax_ids($u)	if $opts{'reassign'};
+$nrows_changed += nullify_tax_ids($u)	if $opts{'nullify'};
+
+printf("%-57s                   %9d\n",
+	   "TOTAL (assigned, reassigned, or nulled)",
+	   $nrows_changed);
 
 exit(0);
 
@@ -40,13 +81,16 @@ exit(0);
 
 sub execute_sth($$;$) {
   my ($fx,$sth,$inc) = @_;
-  $inc = $palias_id_inc unless defined $inc;
-  for(my $b=$palias_id_min; $b<$palias_id_max; $b+=$inc) {
+  my $nr_tot = 0;
+  $inc = $opts{'update-increment'} unless defined $inc;
+  for(my $b=$opts{'palias-id-min'}; $b<$opts{'palias-id-max'}; $b+=$inc) {
 	my $e = $b+$inc;
 	my $nr = $sth->execute($b,$e);
-	printf("\r$fx(;$inc) [%9d,%9d) %9d", $b, $e, $nr);
+	$nr_tot += $nr;
+	printf("\r%-30s [%9d,%9d)    rows:%9d  tot:%9d", "$fx(;$inc)", $b, $e, $nr, $nr_tot);
   }
   print("... done\n");
+  return $nr_tot;
 }
 
 
@@ -54,7 +98,7 @@ sub execute_sth($$;$) {
 sub create_paotax ($) {
   ## create a temp table with palias_id, current tax_id, inferred tax_id
   my $u = shift;
-  eval { $u->do('drop table paotax') };
+  eval { $u->do('drop table paotax') } if $opts{'drop-paotax'};
   $u->do('create table paotax (palias_id integer not null, tax_id integer, infer_tax_id integer) without oids');
   my $sql = qq/
 	INSERT INTO PAOTAX
@@ -64,7 +108,7 @@ sub create_paotax ($) {
     WHERE palias_id>=? and palias_id<?
 	/;
   my $sth = $u->prepare( $sql );
-  execute_sth((caller(0))[3], $sth);
+  execute_sth((caller(0))[3], $sth, $opts{'increment'});
   $u->do('create index paotax_palias_id on paotax(palias_id)');
   $u->do('analyze paotax');
 }
@@ -80,7 +124,7 @@ sub assign_tax_ids ($) {
                         tax_id IS NULL AND infer_tax_id IS NOT NULL);
 	/;
   my $sth = $u->prepare($sql);
-  execute_sth((caller(0))[3], $sth, 100);
+  execute_sth((caller(0))[3], $sth, $opts{'update-increment'});
 }
 
 
@@ -94,7 +138,7 @@ sub reassign_tax_ids ($) {
                         tax_id != infer_tax_id);
 	/;
   my $sth = $u->prepare($sql);
-  execute_sth((caller(0))[3], $sth, 100);
+  execute_sth((caller(0))[3], $sth, $opts{'update-increment'});
 }
 
 
@@ -109,5 +153,5 @@ sub nullify_tax_ids ($) {
                         tax_id IS NOT NULL AND infer_tax_id IS NULL);
 	/;
   my $sth = $u->prepare($sql);
-  execute_sth((caller(0))[3], $sth, 100);
+  execute_sth((caller(0))[3], $sth, $opts{'update-increment'});
 }
